% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeConcordance.R
\name{computeConcordance}
\alias{computeConcordance}
\title{Compute consistency between correlations of cell line sensitivity measures}
\usage{
computeConcordance(
  allCorrelations,
  subsettedCorrelations,
  sensMeasure,
  coefName
)
}
\arguments{
\item{allCorrelations}{A list of dataframes that contain the correlation
coefficients across drugs for all
cell lines, as returned by computeDrugCorrelation().}

\item{subsettedCorrelations}{A list of dataframes that contain the
correlation coefficients across drugs
for only subsetted cell lines, as returned by computeDrugCorrelation().}

\item{sensMeasure}{A character vector specifying the sensitivity
measure of interest. Must be one of the measures included in
correlations parameter.}

\item{coefName}{A character vector specifying the correlation coefficient
of interest. Must be one of the coefficients included in the
correlations parameter.}
}
\value{
Returns an S3 concordance object outlining concordance between
    the coefName correlations of sensMeasure in allCorrelations and in
    subsettedCorrelations.
}
\description{
Computes Harrell's Concordance Index between drug sensitivity measures
across PharmacoSets after removing inconsistent cell lines.
}
\author{
{Casey Hon, \email{casey.hon@mail.utoronto.ca}}
}
