% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeCorrelation.R
\name{computeCorrelation}
\alias{computeCorrelation}
\title{Compute correlations between drug sensitivity measures of cell lines}
\usage{
computeCorrelation(pSet, coefs, pval)
}
\arguments{
\item{pSet}{Intersected PharmacoSet object containing common drugs and/or
cell lines, as returned by PharmacoGx::intersectPSet(). PharmacoSets
should have sensitivity information which can be checked using
PharmacoGx::availablePSets() in the 'type' column.}

\item{coefs}{List of correlation coefficients of interest. Current options
are: "pearson", "spearman", "kendall".}

\item{pval}{Logical of whether or not to include p-values of each
correlation coefficient. Default is TRUE.}

\item{sensMeasures}{List of drug sensitivity measures of interest. Options
depend on the PharmacoSets chosen, and can be identified by looking at
names(pSetName@sensitivity[["profiles"]])}
}
\value{
Returns a list of dataframes, one dataframe for each sensitivity
    measure of interest, where rows correspond to drugs and columns
    correspond to each type of correlation coefficient from coefs.
    Also includes p-values in columns is pval was set to TRUE.
}
\description{
Computes Pearson and Spearman Correlation Coefficients for specified
    drug sensitivity measures (such as IC50 of AUC) and specified
    subset of cell lines
}
\references{
Smirnov P, Safikhani Z, El-Hachem N, Wang D, She A, Olsen C, Freeman M,
    Selby H, Gendoo D, Grossman P, Beck A, Aerts H, Lupien M, Haibe-Kains
    AG, (2016). “PharmacoGx: an R package for analysis of large
    pharmacogenomic datasets.” \emph{Bioinformatics (Oxford, England)}.
}
\author{
{Casey Hon, \email{casey.hon@mail.utoronto.ca}}
}
