% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConsistentCellLines.R
\name{getConsistentCellLines}
\alias{getConsistentCellLines}
\title{Get list of consistent cell lines}
\usage{
getConsistentCellLines(correlations, sensMeasure, coefName, min = 0.5)
}
\arguments{
\item{correlations}{A list of dataframes that contain the correlation
coefficients of the sensitivity measures of interest, as returned
by computeCorrelation().}

\item{sensMeasure}{A character vector defining the sensitivity measure
of interest. Must be one of the measures included in the correlations
dataframe.}

\item{coefName}{A character vector defining the correlation coefficient
of interest. Must be one of the coefficients included in correlations
dataframe.}

\item{min}{A numeric value representing the minimum correlation coefficient
desired. All correlations greater than or equal to min are deemed
consistent. Default is 0.5.}
}
\value{
A dataframe containing only correlations of consistent cell lines.
}
\description{
Identifies cell lines that are deemed consistent based on the
    user-specified minimum correlation of sensitivity measures
    across PharmacoSets.
}
\examples{
# Intersect PharmacoSets of interest based on common cell lines
CTRP <- PharmacoGx::downloadPSet("CTRPv2_2015")
GRAY <- PharmacoGx::downloadPSet("GRAY_2013")
intersected <- PharmacoGx::intersectPSet(c(CTRP, GRAY), intersectOn = c("drugs", "cell.lines"))
correlations <- computeCellLineCorrelation(pSet = intersected,
    coefs = "pearson",
    sensMeasures = "aac_recomputed",
    pval = TRUE)
consistentLines <- getConsistentCellLines(correlations,
    sensMeasure = "aac_recomputed_corrs",
    coefName = "pearson")

}
\author{
{Casey Hon, \email{casey.hon@mail.utoronto.ca}}
}
