% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersectSensMeasures.R
\name{intersectSensMeasures}
\alias{intersectSensMeasures}
\title{Intersect PharmacoSets on common sensitivity measures}
\usage{
intersectSensMeasures(pSet)
}
\arguments{
\item{pSet}{intersected PharmacoSets of interest}
}
\value{
Returns a list of sensitivity measures in common.
}
\description{
Given an object of intersected PharmacoSets, the function finds sensitivity
measures in common.
}
\author{
{Casey Hon, \email{casey.hon@mail.utoronto.ca}}
}
