% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCorrelations.R
\name{plotCorrelations}
\alias{plotCorrelations}
\title{Plot correlation coefficients in bar plots}
\usage{
plotCorrelations(correlations, coefficient, title)
}
\arguments{
\item{correlations}{A dataframe that contains correlation
coefficients.}

\item{coefficient}{A character vector specifying which
correlation coefficient to plot. Must be one of the coefficients
in correlations, as returned by names(correlations).}

\item{title}{A character vector specifying the desired plot title.}
}
\value{
A bar plot with coefficient values on the y-axis, and either
    drugs or cell lines on the x-axis depending on the correlation
    dataframe inputted.
}
\description{
Plots correlation coefficients in bar plot from given
    dataframe, as returned by computeCellLineCorrelation
    or computeDrugCorrelation.
}
\examples{
# Intersect PharmacoSets of interest based on common cell lines
CTRP <- PharmacoGx::downloadPSet("CTRPv2_2015")
GRAY <- PharmacoGx::downloadPSet("GRAY_2013")
intersected <- PharmacoGx::intersectPSet(c(CTRP, GRAY),
    intersectOn = c("drugs", "cell.lines"))
correlations <- computeCorrelation(pSet = intersected,
    coefs = "pearson",
    sensMeasures = "aac_recomputed",
    pval = TRUE)
plotCorrelations(correlations = correlations$aac_recomputed_corrs,
    coefficient = "pearson",
    title = "Pearson Correlations of Recomputed AAC Values")

}
