% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCellLineConsistency.R
\name{runCellLineConsistency}
\alias{runCellLineConsistency}
\title{Run Shiny App for cellLineConsistency}
\usage{
runCellLineConsistency()
}
\value{
No return value, opens up a Shiny page.
}
\description{
A function that launches the Shiny app for cellLineConsistency.
}
\examples{
\dontrun{

cellLineConsistency::runCellLineConsistency()
}

}
\references{
Wickham, H. (2020). Mastering Shiny. O'Reilly Media. \href{https://mastering-shiny.org/basic-app.html}{Link}
}
