% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConsistentCellLines.R
\name{getConsistentCellLines}
\alias{getConsistentCellLines}
\title{Get list of consistent cell lines}
\usage{
getConsistentCellLines(correlations, sensMeasure, coefName, min)
}
\arguments{
\item{correlations}{A list of dataframes that contain the correlation
coefficients of the sensitivity measures of interest, as returned
by computeCorrelation().}

\item{sensMeasure}{A character vector defining the sensitivity measure
of interest. Must be one of the measures included in the correlations
dataframe.}

\item{coefName}{A character vector defining the correlation coefficient
of interest. Must be one of the coefficients included in correlations
dataframe.}

\item{min}{A numeric value representing the minimum correlation coefficient
desired. Default is 0.5.}
}
\value{
A dataframe containing only correlations of consistent cell lines.
}
\description{
Identifies cell lines that are deemed consistent based on the
    user-specified minimum correlation of sensitivity measures
    across PharmacoSets.
}
\author{
{Casey Hon, \email{casey.hon@mail.utoronto.ca}}
}
